requirements.txt
scikit-learn
numpy
mediapipe
opencv-python
gTTS
pygame
Pillow

